Getview update
