+++ lib/widgets/scribble/scribble_model.dart
@@ -187,7 +187,7 @@ class ScribbleModel extends FormFieldModel implements IFormField, IViewableWidge
     super.dispose();
   }
 
-  Widget getView({Key? key}) => ScribbleView(this);
+  Widget getView({Key? key}) => getAnimatedView(ScribbleView(this));
 }
 
 class SignatureModel extends ViewableWidgetModel
